rename : make clean code in execute section #38
