+++ jaemjeon_jisookim_minishell/main/include/execute.h
@@ -6,7 +6,7 @@
 /*   By: jisookim <jisookim@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/01 14:12:35 by jisookim          #+#    #+#             */
-/*   Updated: 2022/09/09 14:34:46 by jisookim         ###   ########.fr       */
+/*   Updated: 2022/09/09 14:58:22 by jisookim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -102,30 +102,22 @@ typedef struct s_exec
 int		count_key_value(t_exec *exec);
 void	make_env_double_ptr(t_exec *exec);
 
-/*
-//init_final_path
-char	*get_paths_from_env(t_exec *exec, char *path_list);
-char	**ft_split_paths(t_exec *exec, char *path_list, char **temp_path_lists);
-char	*get_final_path(t_exec *exec, char **temp_path_lists);
-void	main_get_final_paths(t_exec *exec);
-*/
-
 //exec_init_get_cmd
 void	get_token_count(t_exec *exec);
 int		get_simple_cmd_count(t_exec *exec, int i);
-void	make_ptr_execve_cmds(t_exec *exec);
-char	**init_execve_cmds(t_exec *exec);
+
+//init_struct
+void	set_exec_struct_final_cmd_loop(t_exec *exec, char *cmd_str, int i);
+void	set_exec_process_number_init(t_exec *exec, char *cmd_str, int j);
+void	set_exec_struct_final_cmd_str(t_exec *exec, int j);
+char	*set_final_path_str(t_exec *exec);
+int		init_exec_struct(t_exec *exec, int j);
 
 //init
 int		count_process(t_exec *exec);
 void	make_path_list(t_exec *exec);
 t_exec	*main_init_exec(t_exec *exec, t_cmd *cmd, t_envlst *env, char *envp[]);
 
-void	set_exec_process_number_init(t_exec *exec, char *cmd_str, int j);
-void	set_exec_struct_final_cmd_loop(t_exec *exec, char *cmd_str, int i);
-void	set_exec_struct_final_cmd_str(t_exec *exec, int j);
-char	*set_final_path_str(t_exec *exec);
-int		init_exec_struct(t_exec *exec, int j);
 
 /*
 	MAIN PART
@@ -135,13 +127,6 @@ int		init_exec_struct(t_exec *exec, int j);
 int		check_built_in(t_exec *exec);
 void	exec_go_built_in(t_exec *exec);
 
-//fork
-
-
-//dup2
-int		single_pipe_dup2(t_exec *exec);
-int		multi_pipe_dup2(t_exec *exec);
-
 //main
 void	exec_executing(t_exec *exec, int process_number, int stat);
 int		exec_single_cmd(t_exec *exec);
@@ -149,31 +134,22 @@ int		exec_multi_cmd(t_exec *exec);
 int		execute(t_cmd *cmd, t_envlst *env, char **envp);
 
 //multi_cmd
-
 pid_t	exec_multi_first(t_exec *exec, int i, pid_t *pid);
 pid_t	exec_multi_middle(t_exec *exec, int i, pid_t *pid);
 pid_t	exec_multi_last(t_exec *exec, int i, pid_t *pid);
 
 void	init_pipe_before_exec(t_exec *exec, int i);
 int		multi_process_exceve(t_exec *exec);
-int		exec_multi_child_process(t_exec *exec);
-
-//pipe
-int		exec_pipe_control(t_exec *exec);
-
-//single_cmd
 
 //exec_tools_file
 int		ft_open(const char *filename, int flags);
 int		ft_close(int fd);
 
 //exec_tools
-void	ft_double_free(char **list);
-int		*ft_pipe(int *pipe_fd);
-int		ft_dup2(int fd1, int fd2);
 pid_t	ft_fork(void);
-int		ft_exceve(const char *filename, char *const argv[], char *const envp[]);
+int		ft_dup2(int fd1, int fd2);
 pid_t	ft_wait(int *statloc, int i);
-
+int		*ft_pipe(int *pipe_fd);
+int		ft_exceve(const char *filename, char *const argv[], char *const envp[]);
 
 #endif
