+++ jaemjeon_jisookim_minishell/main/include/utils.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   utils.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jaemjeon <jaemjeon@student.42seoul.kr>     +#+  +:+       +#+        */
+/*   By: jisookim <jisookim@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/24 04:04:48 by jaemjeon          #+#    #+#             */
-/*   Updated: 2022/08/29 04:06:42 by jaemjeon         ###   ########.fr       */
+/*   Updated: 2022/09/09 14:49:06 by jisookim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,5 +21,8 @@ int	ft_has_ifs(char *string);
 int	ft_is_ifs(const char *letter);
 int	ft_skip_ifs1(char **string_pointer);
 void	ft_skip_ifs2(char *line, int *index);
+
+//util_tools
+void	ft_double_free(char **list);
 
 #endif
