+++ jaemjeon_jisookim_minishell/main/src/exe_init/exec_init.c
@@ -6,7 +6,7 @@
 /*   By: jisookim <jisookim@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/02 15:24:31 by jisookim          #+#    #+#             */
-/*   Updated: 2022/09/09 12:18:07 by jisookim         ###   ########.fr       */
+/*   Updated: 2022/09/09 14:58:11 by jisookim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -51,9 +51,6 @@ void	make_path_list(t_exec *exec)
 	return ;
 }
 
-// all of the cmd string
-// main cmd
-// execve(const char *path, char *const argv[], char *const envp[]);
 t_exec	*main_init_exec(t_exec *exec, t_cmd *cmd, t_envlst *env, char **envp)
 {
 	exec = ft_calloc(1, sizeof(t_exec));
@@ -72,116 +69,4 @@ t_exec	*main_init_exec(t_exec *exec, t_cmd *cmd, t_envlst *env, char **envp)
 	make_env_double_ptr(exec);
 	make_path_list(exec);
 	return (exec);
-}
-
-
-// ======================================================
-
-void	set_exec_struct_final_cmd_loop(t_exec *exec, char *cmd_str, int i)
-{
-	cmd_str = ft_strdup(exec->cmds->simple_cmd->string_value);
-	exec->final_cmd_str[i] = ft_strdup(cmd_str); 
-	free(cmd_str);
-	return ;
-}
-
-void	set_exec_process_number_init(t_exec *exec, char *cmd_str, int j)
-{
-	int	i;
-	int	process_number;
-
-	i = 0;
-	process_number = j;
-	exec->cmds = exec->cmd_head;
-	while (i < process_number)
-	{
-		exec->cmds = exec->cmds->next;
-		i++;
-	}
-	;
-}
-
-void	set_exec_struct_final_cmd_str(t_exec *exec, int j) // j == process number
-{
-	int		i;
-	char	*cmd_str;
-
-	exec->final_cmd_str = ft_calloc(1, sizeof(char *) * (exec->token_cnt[j] + 1));
-	exec->final_cmd_str[exec->token_cnt[j]] = NULL;
-	if (!exec->final_cmd_str)
-		ft_exit(exec);
-	
-	set_exec_process_number_init(exec, cmd_str, j);
-	i = 0;
-	while (exec->cmds->simple_cmd->string_value)
-	{
-		set_exec_struct_final_cmd_loop(exec, cmd_str, i);
-		if (exec->cmds->simple_cmd->next)
-			exec->cmds->simple_cmd = exec->cmds->simple_cmd->next;
-		else
-			break;
-		i++;
-	}
-	set_exec_struct_final_cmd_loop(exec, cmd_str, i);
-	exec->cmds = exec->cmd_head; // init
-	return ;
-}
-
-
-char	*set_final_path_str(t_exec *exec)
-{
-	char			*slash_cmd;
-	char			*temp;
-	struct stat		buf;
-	int				i;
-
-	slash_cmd = ft_strjoin("/", exec->final_cmd_str[0]);
-	if (!(slash_cmd))
-		ft_exit(exec);
-	i = 0;
-	while (exec->path_lst[i])
-	{
-		temp = (char *)ft_calloc(1, ft_strlen(exec->path_lst[i]) \
-									+ ft_strlen(slash_cmd) + 1);
-		temp = ft_strjoin(exec->path_lst[i], slash_cmd);
-		if (!temp)
-			ft_exit(exec);
-		if (!stat(temp, &buf))
-		{
-			exec->final_path = temp;
-			return (exec->final_path);
-		}
-		free(temp);
-		i++;
-	}
-	free(slash_cmd);
-	return (NULL);
-}
-
-
-int	init_exec_struct(t_exec *exec, int process_number)
-{
-	char *temp;
-	//check redirection
-	
-	set_exec_struct_final_cmd_str(exec, process_number);
-
-	// // debug
-	// dprintf(2, "\n\n=============[%d] EXEC DEBUG ============\n", process_number);
-	// int idx = 0;
-	// dprintf(2, "process_number : %d\n", process_number);
-	// dprintf(2, "exec->token_cnt[process_number]: %d\n", exec->token_cnt[process_number]);
-	
-	// while (idx < exec->token_cnt[process_number]) 
-	// {
-	// 	dprintf(2, "exec->final_cmd_str[%d] : %s\n", idx, exec->final_cmd_str[idx]);
-	// 	idx++;
-	// }
-	// dprintf(2, "exec->final_cmd_str[%d] : %s (last)\n", idx, exec->final_cmd_str[idx]); // needs to have (null);
-	// dprintf(2, "=============EXEC DEBUG ============\n\n");
-	// // debug
-
-	exec->final_path = set_final_path_str(exec);
-
-	return (0);
 }
