+++ jaemjeon_jisookim_minishell/main/src/exe_init/exec_init_get_cmd.c
@@ -6,7 +6,7 @@
 /*   By: jisookim <jisookim@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/03 14:55:52 by jisookim          #+#    #+#             */
-/*   Updated: 2022/09/06 01:04:06 by jisookim         ###   ########.fr       */
+/*   Updated: 2022/09/09 14:57:03 by jisookim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
