--- jaemjeon_jisookim_minishell/main/src/execute/exec_fork.c
@@ -1,15 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   exec_fork.c                                        :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: jisookim <jisookim@student.42seoul.kr>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/09/01 15:15:47 by jisookim          #+#    #+#             */
-/*   Updated: 2022/09/07 21:05:05 by jisookim         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "../../include/minishell.h"
-
-// divide handler depending parameter 'i' - first / middle / last
