+++ jaemjeon_jisookim_minishell/main/src/utils/util_tools.c
@@ -1,13 +1,26 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   exec_single_cmd.c                                  :+:      :+:    :+:   */
+/*   util_tools.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: jisookim <jisookim@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/09/07 18:21:33 by jisookim          #+#    #+#             */
-/*   Updated: 2022/09/07 20:56:21 by jisookim         ###   ########.fr       */
+/*   Created: 2022/09/09 14:48:26 by jisookim          #+#    #+#             */
+/*   Updated: 2022/09/09 14:48:42 by jisookim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../include/minishell.h"
+
+void	ft_double_free(char **list)
+{
+	int	i;
+
+	i = 0;
+	while (list[i])
+	{
+		free(list[i]);
+		i++;
+	}
+	free(list);
+}
